pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bio-mimetic-swarm"
version = "2.0.0"
description = "Decentralized swarm intelligence using Lennard-Jones artificial potential fields"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Bio-Mimetic Swarm Project"},
]
keywords = [
    "swarm-robotics", "multi-agent-systems", "lennard-jones",
    "artificial-potential-fields", "decentralized-control",
    "autonomous-surface-vessels", "bio-mimetic", "simulation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "scipy>=1.11.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
scenarios = ["pyyaml>=6.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/your-handle/bio-mimetic-swarm"
Repository = "https://github.com/your-handle/bio-mimetic-swarm"
Documentation = "https://github.com/your-handle/bio-mimetic-swarm/tree/main/docs"
Issues = "https://github.com/your-handle/bio-mimetic-swarm/issues"

[project.scripts]
swarm = "swarm.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
swarm = ["py.typed"]

# ── Ruff ─────────────────────────────────────────────────────────────────────

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pyflakes + pycodestyle
    "I",              # isort
    "N",              # pep8-naming
    "UP",             # pyupgrade
    "B",              # flake8-bugbear
    "SIM",            # flake8-simplify
    "RUF",            # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["swarm"]

# ── Mypy ─────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
packages = ["swarm"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["scipy.*", "matplotlib.*"]
ignore_missing_imports = true

# ── Pytest ───────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests that take >5s (deselect with '-m \"not slow\"')",
]

# ── Coverage ─────────────────────────────────────────────────────────────────

[tool.coverage.run]
source = ["swarm"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
